# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/user/info/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户列表",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "业务API"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户详情",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.User": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "astrology": {
                    "description": "星座",
                    "type": "string"
                },
                "avatar": {
                    "description": "小头像",
                    "type": "string"
                },
                "create_time": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "medium": {
                    "description": "大头像",
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "passport": {
                    "description": "用户密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号码",
                    "type": "integer"
                },
                "sex": {
                    "description": "性别:1男2女",
                    "type": "integer"
                },
                "source": {
                    "description": "来源 0.真实用户 1.糗事百科 2.段子手",
                    "type": "string"
                },
                "status": {
                    "description": "1有效 2无效",
                    "type": "integer"
                },
                "update_time": {
                    "description": "更新时间",
                    "type": "integer"
                },
                "username": {
                    "description": "用户姓名",
                    "type": "string"
                }
            }
        }
    }
}